fix upload photo
