+++ offlines/characters/pokemon/Bulbasaur.json
@@ -20,7 +20,7 @@
   "IdName": "Bulbasaur_#1",
   "Level": 1,
   "Name": "Bulbasaur",
-  "Photo": "hinata.jpg",
+  "Photo": "arcko.png",
   "Rank": "Common",
   "Shape": "",
   "Short name": "Bulbasaur",
